[doc][patch] restructure code
