+++ .changeset/hot-cows-exercise.md
@@ -1,0 +1,5 @@
+---
+"@nil-/doc": patch
+---
+
+[doc][patch] restructure code
