+++ projects/doc/src/lib/components/layout/VerticalPanel.svelte
