+++ projects/doc/src/lib/index.ts
@@ -10,7 +10,7 @@ export type {
     Unionized
 } from "./components/block/controls/types";
 
-export { default as Layout } from "./components/Layout.svelte";
+export { default as Layout } from "./components/layout/Layout.svelte";
 export { default as Instance } from "./components/block/Instance.svelte";
 export { default as Block } from "./components/block/Block.svelte";
 export { default as Template } from "./components/block/Template.svelte";
